+++ backend/canisters/community/impl/src/model/user_event_batch.rs
@@ -1,41 +1,10 @@
 use crate::{can_borrow_state, run_regular_jobs};
-use timer_job_queues::{TimerJobItem, TimerJobItemGroup};
+use timer_job_queues::{grouped_timer_job_batch, TimerJobItem, TimerJobItemGroup};
 use types::UserId;
 use user_canister::CommunityCanisterEvent;
 use utils::canister::should_retry_failed_c2c_call;
 
-pub struct UserEventBatch {
-    user_id: UserId,
-    events: Vec<CommunityCanisterEvent>,
-}
-
-impl TimerJobItemGroup for UserEventBatch {
-    type Key = UserId;
-    type Item = CommunityCanisterEvent;
-
-    fn new(user_id: UserId) -> Self {
-        UserEventBatch {
-            user_id,
-            events: Vec::new(),
-        }
-    }
-
-    fn key(&self) -> UserId {
-        self.user_id
-    }
-
-    fn add(&mut self, event: CommunityCanisterEvent) {
-        self.events.push(event)
-    }
-
-    fn into_items(self) -> Vec<CommunityCanisterEvent> {
-        self.events
-    }
-
-    fn is_full(&self) -> bool {
-        self.events.len() >= 1000
-    }
-}
+grouped_timer_job_batch!(UserEventBatch, UserId, CommunityCanisterEvent, 1000);
 
 impl TimerJobItem for UserEventBatch {
     async fn process(&self) -> Result<(), bool> {
@@ -44,9 +13,9 @@ impl TimerJobItem for UserEventBatch {
         }
 
         let response = user_canister_c2c_client::c2c_notify_community_canister_events(
-            self.user_id.into(),
+            self.key.into(),
             &user_canister::c2c_notify_community_canister_events::Args {
-                events: self.events.clone(),
+                events: self.items.clone(),
             },
         )
         .await;